fix long->i32

+++ sys/src/cmd/usb/disk/ums.h
@@ -86,8 +86,8 @@ struct Ums
 struct Cbw
 {
 	char	signature[4];		/* "USBC" */
-	long	tag;
-	long	datalen;
+	i32	tag;
+	i32	datalen;
 	u8	flags;
 	u8	lun;
 	u8	len;
@@ -97,8 +97,8 @@ struct Cbw
 struct Csw
 {
 	char	signature[4];		/* "USBS" */
-	long	tag;
-	long	dataresidue;
+	i32	tag;
+	i32	dataresidue;
 	u8	status;
 };
